LayoutTests/platform/wpe/fullscreen/exit-full-screen-video-crash-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: InvalidStateError: The object is in an invalid state.
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Type error
